Refactor environment variable loading and database initialization
